Add sanity check for elfsymbol value offset (#104)
