+++ ddprof-lib/src/test/resources/readme.txt
@@ -1,0 +1,3 @@
+TODO!
+Create a library exhibiting the problems with the st_value offsets in the ELF file format.
+Put it here and update the elfparser_ut.cpp test to use it.
